FIX-0004 - refactor packages for project
