+++ app/src/main/java/ru/easycode/words504/languages/data/cache/ChosenLanguageCache.kt
@@ -3,8 +3,8 @@ package ru.easycode.words504.languages.data.cache
 import ru.easycode.words504.data.cache.storage.ObjectStorage
 
 interface ChosenLanguageCache {
-    interface Read : ru.easycode.words504.core.Read<LanguageCache>
-    interface Save : ru.easycode.words504.core.Save<LanguageCache>
+    interface Read : ru.easycode.words504.data.Read<LanguageCache>
+    interface Save : ru.easycode.words504.data.Save<LanguageCache>
     interface Mutable : Read, Save
 
     class Base(
