+++ app/src/main/java/ru/easycode/words504/languages/data/cache/LanguageCache.kt
@@ -1,6 +1,6 @@
 package ru.easycode.words504.languages.data.cache
 
-import ru.easycode.words504.core.Empty
+import ru.easycode.words504.data.Empty
 
 interface LanguageCache : Empty {
     data class Base(private val key: String, private val name: String) : LanguageCache {
