+++ app/src/main/java/ru/easycode/words504/languages/domain/HandleHttpError.kt
@@ -1,7 +1,7 @@
 package ru.easycode.words504.languages.domain
 
 import retrofit2.Response
-import ru.easycode.words504.core.HandleError
+import ru.easycode.words504.domain.HandleError
 
 class HandleHttpError : HandleError<Response<*>, Throwable> {
