+++ app/src/main/java/ru/easycode/words504/tts/presentation/TTSCommunication.kt
@@ -1,4 +1,4 @@
-package ru.easycode.words504.tts
+package ru.easycode.words504.tts.presentation
 
 import ru.easycode.words504.presentation.Communication
