+++ core/src/main/java/ru/easycode/words504/data/Empty.kt
@@ -1,4 +1,4 @@
-package ru.easycode.words504.core
+package ru.easycode.words504.data
 
 interface Empty {
     fun isEmpty(): Boolean
