+++ core/src/main/java/ru/easycode/words504/domain/HandleError.kt
@@ -1,4 +1,4 @@
-package ru.easycode.words504.core
+package ru.easycode.words504.domain
 
 interface HandleError<S : Any, E : Any> {
     fun handle(source: S): E
