+++ core/src/main/java/ru/easycode/words504/domain/Mapper.kt
@@ -1,4 +1,4 @@
-package ru.easycode.words504.core
+package ru.easycode.words504.domain
 
 interface Mapper<S : Any, R : Any> {
